fix(inputs.jenkins): Filter after searching sub-folders
Currently, the Jenkins plugin will apply the user filter to directory
names immediately. If the directory does not match the filter it would
toss the folder. This prevents traversing into the folder itself where a
user might want a job inside the folder.

This change will move the filtering operation to occur after we kick off
directory traversal. The changes to the tests were necessary since we
never went into the folders correctly. The test results are the same,
the changes are to the response our mock jenkins server provides.

fixes: #14054