fix(processors.starlark): Better handling of tracking metrics
The starlark process is used to make modifications to metrics, create
new metrics, or even remove metrics. When the starlark script returns it
does not return a metric with tracking information, only a generic
metric.

Currently we test to see if the metric returned is the same, but even if
they are idential, the tracking informiation is lost and the metric is
never marked accepted or dropped, hence we lose the tracking
information.

Instead of accepting everything, track the hash ID of each metric.
If the metric is the same, then continue on with the original metric.
For tracking metrics, we accept the new metric, since we can count on the
original metric to get ack'ed or nack'ed by the output.