test(inputs.vault): Avoid hard-compare of metrics
During the integration test we spin up a real vault instance in a docker
container and hit the metrics endpoint. However, the number of metrics
reported by this endpoint change depending on how long the instance has
been up.

For example, the longer then container is up additional metrics about
the go runtime start to show up.

This means doing a hard-comparison of the metrics we expect is not a
good idea and instead we should ensure that we collected without error.
We already have unit tests that validate we can correctly parse the
various types of metrics.