fix(inputs.mqtt_consumer): Resolve could not mark message delivered issue
The lock for the messages map was collected after sending the metric to
the accumulator. This was not the right behavior, as the accumulator
could send the metric before we even got to the next step to lock and
add the message to the map. As a result, in some scenarios the metric
could get sent and attempt to mark it delivered before we even record
the metric.

fix: #14242