feat(outputs.stackdriver): Enable histogram support
This enables the stackdriver output to send histogram metrics. These
metrics are required to have a sum field and all other fields should
include the buckets and counts in a single metric. Prometheus parser
metric version 2 is not supported given the sparse nature of the
metrics.

This also increases the label limit to 30 per the stackdriver published
limits. As histograms could have even more labels than other types of
metrics.

fixes: #14258