test(inputs.vault): Remove metric comparison in integration test
The integration test has once again proved to be flaky. It seems that
depending on when we get metrics during the boot up cycle we may or may
not get the correct metrics we are looking for or additional metrics.

We already check for certain metrics, and that we can parse them in a
unit test, so instead use the integration test to verify that we can
sucessfully connect and eventually return metrics.