fix(processors.starlark): Resolve deepcopy double refcount
The call to copy currently will make a deepcopy of the metric, but
immediately call drop on it as well. This makes no sense as if that
metric ends up in an output, it could still get dropped or accepted.
This can lead to issues where we double count the refcount.

Additionally, the current state of checking for original metrics is also
resulting in a double accept. The problem here is our tests! When we go
to accept the metrics to simulate an output, the metric is no longer a
tracking metric, meaning the accept call is a no-op, and does nothing.

fixes: #14360