fix: #issue 12500 occasional serverFault
resolves #12500

the serverFaultCode "A specified parameter was not correct:
querySpec.startTime, querySpec.endTime" results from requests for
metrics from the future.

Those requests get planned (check: hiwater mark, hwMark) for a
plugin-run (-> plugin "interval") in the future in relation to the
estimated time the vcenter will have collected the next metrics.

This estimated time/interval gets mangled up on a connection error
(e.g. connection reset), because the est. interval gets reset to 1.

Solution: If we realize (ourselves) that the metric we are about to
request is from the future, then... don't request it. Skip it.