Refactor systemd_units for multiple subcommands and add `show` subcommand
This commit refactores the systemd_units input plugin to allow different
subcommands of `systemctl` to be used as inputs. It introduces the
`subcommand` option to switch between the different subcommands and
extends the test-suite to allow testing multiple subcommands.

The test suite was extended to verify the parameter generation for the
`systemctl` command.

Add the `show` subcommand to systemd_units. Using the `show` subcommand
allows gathering more information about the unit. This change is not
backward compatible.

The `subcommand` option allows switching between the `list-units`
subcommand (best compatibility) or the `show` subcommand for more
information.

The default is the `list-units` subcommand. Upgrading does not lead to
a change in behaviour.

See [issue #14521](https://github.com/influxdata/telegraf/issues/14521)
for more details.