fix(inputs.prometheus): Remove duplicate response_timeout option
In #14153, the HTTP client config struct gained a response timeout
config option with TOML tags. This meant that there were two defined for
both Prometheus plugin and the HTTP client config struct. This removes
the one in Prometheus, which was used to set the one in the HTTP client
anyway.

fixes: #15076