fix(plugins.input.redis): Discard invalid errorstat lines (#15311)
Some badly-behaved sources (such as dragonfly) may send error messages
which are not in the expected format. If errorstat lines are found
which do not include an `=` sign, silently discard them.

Somewhere between 1.26 and 1.28.5, a change was made which caused such
missing '=' symbols to cause a panic.

In addition to ignoring such entries, add debug-level logging for this
and a similar problem, where the errorstat values are not base-10.