feat(inputs.statsd): Allow counters to report as float
By default counters are stored as ints. However, if a user is using a
value as both a float or a int it can lead to type conflicts. This
maintains the interal representation as an int to avoid messing with the
caching code, but allows the counter to be reported as a float when a
metric is reported.

fixes: #1978