fix(inputs.postgresql_extensible): Use same timestamp for each gather
Rather than using a different timestamp for each query, this uses a
single timestamp determined at the start of gather, for any and all
metrics generated during that gather.

While the timestamp might be off from when the actual queries are run,
it combines all metrics around a single timestamp to better understand
when data was captured.

fixes: #9737