test(inputs.syslog): Add timeout to closed socket test
This is the most flaky test we currently have. It was omitted from
Windows testing because it was failing consistently and then started
happening on Linux as well.

The Test fails after waiting 9mins waiting to read from the socket. With
no timeout set in the config the connection is expected to stay open
indefinitly as described in #10121.