feat(agent): Enable watching for new files
This allows for watching for new files in a configuraiton directory.
When a new file is found, Telegraf will reload with that file added to
the list of files to watch. Note this will reload when any file, a
telegraf config or not, is added to the directory.

TODO: test on Windows

fixes: #12389
fixes: #12264