fix(plugin,procstat): fix cpu usage percentage with multiple filters
With the "new" style plugin filters, if multiple filters were used, the
list of running processes was being reset on each filter iteration
instead of globally. This commit fixes this by setting the running pid
map outside the filter iteration. The cpu usage percentage calculation
needs multiple passes of gather to calculate the difference so
recreating the proc on each iteration meant it never got past the first
pass.