chore(outputs.sql): Update ClickHouse Docker image for test
The Docker image `yandex/clickhouse-server` that has been used for the
integration test hasn't been updated in 3 years. ClickHouse has moved to
the Docker image that is called `clickhouse`.

This uses the new image for the integration test.

The log messages have changed, so that is taken into account when
waiting for the test container to spin up. Technically waiting for the
open port would be enough but for good measure we also wait for a
corresponding log output. To get the log output, a config file is
included.