feat(ping): Add option to use start timestamp
Since ping can take multiple seconds to finish executing, and it may not
be uncommon to set the interval between pings to a second, provide an
option to calculate the timestamp for the measurement at the start of
the collection, rather than the end.  This may make other comparisons
(jitter, etc) easier in downstream TSDBs.