feat(outputs.sql): Use Prepared statements for write tx
The current approach to writing, issues an INSERT query for every
metric in the buffer, which while working, leaves room for improvement.

This commit modifies the Write behavior as follows:

- Sort the metrics fields/tags (to ensure greater cache-hit ratio)
- Use Prepared statements to write multiple metrics with a single query

The results depends on the metrics buffer, the metric cardinality, the
output server configuration, etc.

That being said, in micro-benchmarks and production deployment,
we've seen 10-100x improvement in write performance.