fix(inputs.opcua): Fix the unit tests
Problems with the current implementation.

Once the client is connected, changing its URL does not trigger an error
because the connection is already established, and the URL is only used
when establishing a new connection.

Reconnecting uses the URL from `o.client.OpcUAClient.Config.Endpoint`,
not from `o.client.Config.OpcUAClientConfig.Endpoint`.