HOS-23177: user profile is not present in ClientStats (#66)
* HOS-23177: user profile is not present in ClientStats

We were facing issue in recieving user-profile
in telegraf since it uses dcd shared memory.
Fixed it now using telegraf helper exe.

* HOS-23203: Added bucket number to noise,snr,rssi

As an workaroud to overcome array sorting issue,
we are adding a bucket number in each bucket of
the nois, snr and rssi array.

* HOS-23214: We need channel field in RfStats

Adding channel in rfstat.